+++ galaxy_utils/__init__.py
@@ -1,9 +1,1 @@
 __version__ = '1.1.5'
-
-
-PROJECT_NAME = "galaxy_sequence_utils"
-PROJECT_OWNER = PROJECT_USERAME = "galaxyproject"
-PROJECT_AUTHOR = 'Dan Blankenberg, Galaxy Project and Community'
-PROJECT_EMAIL = 'galaxy-dev@lists.galaxyproject.org'
-
-PROJECT_URL = "https://github.com/galaxyproject/sequence_utils"